characters-api/app/config.py
import logging
from dotenv import load_dotenv
import os

# Загрузка переменных из .env файла
load_dotenv()

# Получение URL базы данных из переменных окружения
DATABASE_URL = os.getenv("DATABASE_URL")

def setup_logging():
    logging.basicConfig(level=logging.INFO)
    logger = logging.getLogger(__name__)
    return logger
